requirements.txt
pygame-ce>=2.5.0
pytmx>=3.31
opencv-python>=4.8.0
